refactor: Loader를 CircularLoader로 수정 [TRL-148]
