+++ src/components/UserInfo/DynamicUserNickname/index.tsx
@@ -6,8 +6,8 @@ import styled, { css } from 'styled-components';
 import HTTP from '@/api';
 import { ReactComponent as CheckIcon } from '@/assets/check.svg';
 import { ReactComponent as EditIcon } from '@/assets/pencil.svg';
+import CircularLoader from '@/components/common/CircularLoader/index';
 import Description from '@/components/common/Description';
-import CircularLoader from '@/components/common/Loader/index';
 import color from '@/constants/color';
 import useGetUserProfile from '@/queryHooks/useGetUserProfile';
 import { nicknameRegExp } from '@/utils/regExp';
